Struct rusoto_clouddirectory::BatchReadOperation

pub struct BatchReadOperation {
    pub get_object_information: Option<BatchGetObjectInformation>,
    pub list_attached_indices: Option<BatchListAttachedIndices>,
    pub list_incoming_typed_links: Option<BatchListIncomingTypedLinks>,
    pub list_index: Option<BatchListIndex>,
    pub list_object_attributes: Option<BatchListObjectAttributes>,
    pub list_object_children: Option<BatchListObjectChildren>,
    pub list_object_parent_paths: Option<BatchListObjectParentPaths>,
    pub list_object_policies: Option<BatchListObjectPolicies>,
    pub list_outgoing_typed_links: Option<BatchListOutgoingTypedLinks>,
    pub list_policy_attachments: Option<BatchListPolicyAttachments>,
    pub lookup_policy: Option<BatchLookupPolicy>,
}

Represents the output of a BatchRead operation.

Fields
get_object_information: Option<BatchGetObjectInformation>

Retrieves metadata about an object.

list_attached_indices: Option<BatchListAttachedIndices>

Lists indices attached to an object.

list_incoming_typed_links: Option<BatchListIncomingTypedLinks>

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed link.

list_index: Option<BatchListIndex>

Lists objects attached to the specified index.

list_object_attributes: Option<BatchListObjectAttributes>

Lists all attributes that are associated with an object.

list_object_children: Option<BatchListObjectChildren>

Returns a paginated list of child objects that are associated with a given object.

list_object_parent_paths: Option<BatchListObjectParentPaths>

Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects. For more information about objects, see Directory Structure.

list_object_policies: Option<BatchListObjectPolicies>

Returns policies attached to an object in pagination fashion.

list_outgoing_typed_links: Option<BatchListOutgoingTypedLinks>

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed link.

list_policy_attachments: Option<BatchListPolicyAttachments>

Returns all of the ObjectIdentifiers to which a given policy is attached.

lookup_policy: Option<BatchLookupPolicy>

Lists all policies from the root of the Directory to the object specified. If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier, policyId, and policyType. Paths that don't lead to the root from the target object are ignored. For more information, see Policies.