Struct rusoto_clouddirectory::BatchWriteOperationResponse

pub struct BatchWriteOperationResponse {
    pub add_facet_to_object: Option<BatchAddFacetToObjectResponse>,
    pub attach_object: Option<BatchAttachObjectResponse>,
    pub attach_policy: Option<BatchAttachPolicyResponse>,
    pub attach_to_index: Option<BatchAttachToIndexResponse>,
    pub attach_typed_link: Option<BatchAttachTypedLinkResponse>,
    pub create_index: Option<BatchCreateIndexResponse>,
    pub create_object: Option<BatchCreateObjectResponse>,
    pub delete_object: Option<BatchDeleteObjectResponse>,
    pub detach_from_index: Option<BatchDetachFromIndexResponse>,
    pub detach_object: Option<BatchDetachObjectResponse>,
    pub detach_policy: Option<BatchDetachPolicyResponse>,
    pub detach_typed_link: Option<BatchDetachTypedLinkResponse>,
    pub remove_facet_from_object: Option<BatchRemoveFacetFromObjectResponse>,
    pub update_object_attributes: Option<BatchUpdateObjectAttributesResponse>,
}

Represents the output of a BatchWrite response operation.

Fields
add_facet_to_object: Option<BatchAddFacetToObjectResponse>

The result of an add facet to object batch operation.

attach_object: Option<BatchAttachObjectResponse>

Attaches an object to a Directory.

attach_policy: Option<BatchAttachPolicyResponse>

Attaches a policy object to a regular object. An object can have a limited number of attached policies.

attach_to_index: Option<BatchAttachToIndexResponse>

Attaches the specified object to the specified index.

attach_typed_link: Option<BatchAttachTypedLinkResponse>

Attaches a typed link to a specified source and target object. For more information, see Typed link.

create_index: Option<BatchCreateIndexResponse>

Creates an index object. See Indexing for more information.

create_object: Option<BatchCreateObjectResponse>

Creates an object in a Directory.

delete_object: Option<BatchDeleteObjectResponse>

Deletes an object in a Directory.

detach_from_index: Option<BatchDetachFromIndexResponse>

Detaches the specified object from the specified index.

detach_object: Option<BatchDetachObjectResponse>

Detaches an object from a Directory.

detach_policy: Option<BatchDetachPolicyResponse>

Detaches a policy from a Directory.

detach_typed_link: Option<BatchDetachTypedLinkResponse>

Detaches a typed link from a specified source and target object. For more information, see Typed link.

remove_facet_from_object: Option<BatchRemoveFacetFromObjectResponse>

The result of a batch remove facet from object operation.

update_object_attributes: Option<BatchUpdateObjectAttributesResponse>

Updates a given object’s attributes.