Struct rusoto_clouddirectory::CloudDirectoryClient

pub struct CloudDirectoryClient { /* fields omitted */ }

A client for the Amazon CloudDirectory API.

Methods

impl CloudDirectoryClient

pub fn new(region: Region) -> CloudDirectoryClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CloudDirectoryClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,