Crate rusoto_clouddirectory

Amazon Cloud Directory
Amazon Cloud Directory is a component of the AWS Directory Service that simplifies the development and management of cloud-scale web, mobile, and IoT applications. This guide describes the Cloud Directory operations that you can call programmatically and includes detailed information on data types and errors. For information about AWS Directory Services features, see AWS Directory Service and the AWS Directory Service Administration Guide.

If you're using the service, you're probably looking for CloudDirectoryClient and CloudDirectory.

Structs

AddFacetToObjectRequest

AddFacetToObjectResponse

ApplySchemaRequest

ApplySchemaResponse

AttachObjectRequest

AttachObjectResponse

AttachPolicyRequest

AttachPolicyResponse

AttachToIndexRequest

AttachToIndexResponse

AttachTypedLinkRequest

AttachTypedLinkResponse

AttributeKey
A unique identifier for an attribute.

AttributeKeyAndValue
The combination of an attribute key and an attribute value.

AttributeNameAndValue
Identifies the attribute name and value for a typed link.

BatchAddFacetToObject
Represents the output of a batch add facet to object operation.

BatchAddFacetToObjectResponse
The result of a batch add facet to object operation.

BatchAttachObject
Represents the output of an AttachObject operation.

BatchAttachObjectResponse
Represents the output batch AttachObject response operation.

BatchAttachPolicy
Attaches a policy object to a regular object inside a BatchRead operation. For more information, see AttachPolicy and BatchReadRequest$Operations.

BatchAttachPolicyResponse
Represents the output of an AttachPolicy response operation.

BatchAttachToIndex
Attaches the specified object to the specified index inside a BatchRead operation. For more information, see AttachToIndex and BatchReadRequest$Operations.

BatchAttachToIndexResponse
Represents the output of a AttachToIndex response operation.

BatchAttachTypedLink
Attaches a typed link to a specified source and target object inside a BatchRead operation. For more information, see AttachTypedLink and BatchReadRequest$Operations.

BatchAttachTypedLinkResponse
Represents the output of a AttachTypedLink response operation.

BatchCreateIndex
Creates an index object inside of a BatchRead operation. For more information, see CreateIndex and BatchReadRequest$Operations.

BatchCreateIndexResponse
Represents the output of a CreateIndex response operation.

BatchCreateObject
Represents the output of a CreateObject operation.

BatchCreateObjectResponse
Represents the output of a CreateObject response operation.

BatchDeleteObject
Represents the output of a DeleteObject operation.

BatchDeleteObjectResponse
Represents the output of a DeleteObject response operation.

BatchDetachFromIndex
Detaches the specified object from the specified index inside a BatchRead operation. For more information, see DetachFromIndex and BatchReadRequest$Operations.

BatchDetachFromIndexResponse
Represents the output of a DetachFromIndex response operation.

BatchDetachObject
Represents the output of a DetachObject operation.

BatchDetachObjectResponse
Represents the output of a DetachObject response operation.

BatchDetachPolicy
Detaches the specified policy from the specified directory inside a BatchWrite operation. For more information, see DetachPolicy and BatchWriteRequest$Operations.

BatchDetachPolicyResponse
Represents the output of a DetachPolicy response operation.

BatchDetachTypedLink
Detaches a typed link from a specified source and target object inside a BatchRead operation. For more information, see DetachTypedLink and BatchReadRequest$Operations.

BatchDetachTypedLinkResponse
Represents the output of a DetachTypedLink response operation.

BatchGetLinkAttributes
Retrieves attributes that are associated with a typed link inside a BatchRead operation. For more information, see GetLinkAttributes and BatchReadRequest$Operations.

BatchGetLinkAttributesResponse
Represents the output of a GetLinkAttributes response operation.

BatchGetObjectAttributes
Retrieves attributes within a facet that are associated with an object inside an BatchRead operation. For more information, see GetObjectAttributes and BatchReadRequest$Operations.

BatchGetObjectAttributesResponse
Represents the output of a GetObjectAttributes response operation.

BatchGetObjectInformation
Retrieves metadata about an object inside a BatchRead operation. For more information, see GetObjectInformation and BatchReadRequest$Operations.

BatchGetObjectInformationResponse
Represents the output of a GetObjectInformation response operation.

BatchListAttachedIndices
Lists indices attached to an object inside a BatchRead operation. For more information, see ListAttachedIndices and BatchReadRequest$Operations.

BatchListAttachedIndicesResponse
Represents the output of a ListAttachedIndices response operation.

BatchListIncomingTypedLinks
Returns a paginated list of all the incoming TypedLinkSpecifier information for an object inside a BatchRead operation. For more information, see ListIncomingTypedLinks and BatchReadRequest$Operations.

BatchListIncomingTypedLinksResponse
Represents the output of a ListIncomingTypedLinks response operation.

BatchListIndex
Lists objects attached to the specified index inside a BatchRead operation. For more information, see ListIndex and BatchReadRequest$Operations.

BatchListIndexResponse
Represents the output of a ListIndex response operation.

BatchListObjectAttributes
Represents the output of a ListObjectAttributes operation.

BatchListObjectAttributesResponse
Represents the output of a ListObjectAttributes response operation.

BatchListObjectChildren
Represents the output of a ListObjectChildren operation.

BatchListObjectChildrenResponse
Represents the output of a ListObjectChildren response operation.

BatchListObjectParentPaths
Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects inside a BatchRead operation. For more information, see ListObjectParentPaths and BatchReadRequest$Operations.

BatchListObjectParentPathsResponse
Represents the output of a ListObjectParentPaths response operation.

BatchListObjectPolicies
Returns policies attached to an object in pagination fashion inside a BatchRead operation. For more information, see ListObjectPolicies and BatchReadRequest$Operations.

BatchListObjectPoliciesResponse
Represents the output of a ListObjectPolicies response operation.

BatchListOutgoingTypedLinks
Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object inside a BatchRead operation. For more information, see ListOutgoingTypedLinks and BatchReadRequest$Operations.

BatchListOutgoingTypedLinksResponse
Represents the output of a ListOutgoingTypedLinks response operation.

BatchListPolicyAttachments
Returns all of the ObjectIdentifiers to which a given policy is attached inside a BatchRead operation. For more information, see ListPolicyAttachments and BatchReadRequest$Operations.

BatchListPolicyAttachmentsResponse
Represents the output of a ListPolicyAttachments response operation.

BatchLookupPolicy
Lists all policies from the root of the Directory to the object specified inside a BatchRead operation. For more information, see LookupPolicy and BatchReadRequest$Operations.

BatchLookupPolicyResponse
Represents the output of a LookupPolicy response operation.

BatchReadException
The batch read exception structure, which contains the exception type and message.

BatchReadOperation
Represents the output of a BatchRead operation.

BatchReadOperationResponse
Represents the output of a BatchRead response operation.

BatchReadRequest

BatchReadResponse

BatchReadSuccessfulResponse
Represents the output of a BatchRead success response operation.

BatchRemoveFacetFromObject
A batch operation to remove a facet from an object.

BatchRemoveFacetFromObjectResponse
An empty result that represents success.

BatchUpdateLinkAttributes
Updates a given typed link’s attributes inside a BatchRead operation. Attributes to be updated must not contribute to the typed link’s identity, as defined by its IdentityAttributeOrder. For more information, see UpdateLinkAttributes and BatchReadRequest$Operations.

BatchUpdateLinkAttributesResponse
Represents the output of a UpdateLinkAttributes response operation.

BatchUpdateObjectAttributes
Represents the output of a BatchUpdate operation.

BatchUpdateObjectAttributesResponse
Represents the output of a BatchUpdate response operation.

BatchWriteOperation
Represents the output of a BatchWrite operation.

BatchWriteOperationResponse
Represents the output of a BatchWrite response operation.

BatchWriteRequest

BatchWriteResponse

CloudDirectoryClient
A client for the Amazon CloudDirectory API.

CreateDirectoryRequest

CreateDirectoryResponse

CreateFacetRequest

CreateFacetResponse

CreateIndexRequest

CreateIndexResponse

CreateObjectRequest

CreateObjectResponse

CreateSchemaRequest

CreateSchemaResponse

CreateTypedLinkFacetRequest

CreateTypedLinkFacetResponse

DeleteDirectoryRequest

DeleteDirectoryResponse

DeleteFacetRequest

DeleteFacetResponse

DeleteObjectRequest

DeleteObjectResponse

DeleteSchemaRequest

DeleteSchemaResponse

DeleteTypedLinkFacetRequest

DeleteTypedLinkFacetResponse

DetachFromIndexRequest

DetachFromIndexResponse

DetachObjectRequest

DetachObjectResponse

DetachPolicyRequest

DetachPolicyResponse

DetachTypedLinkRequest

Directory
Directory structure that includes the directory name and directory ARN.

DisableDirectoryRequest

DisableDirectoryResponse

EnableDirectoryRequest

EnableDirectoryResponse

Facet
A structure that contains Name, ARN, Attributes, Rules, and ObjectTypes. See Facets for more information.

FacetAttribute
An attribute that is associated with the Facet.

FacetAttributeDefinition
A facet attribute definition. See Attribute References for more information.

FacetAttributeReference
The facet attribute reference that specifies the attribute definition that contains the attribute facet name and attribute name.

FacetAttributeUpdate
A structure that contains information used to update an attribute.

GetAppliedSchemaVersionRequest

GetAppliedSchemaVersionResponse

GetDirectoryRequest

GetDirectoryResponse

GetFacetRequest

GetFacetResponse

GetLinkAttributesRequest

GetLinkAttributesResponse

GetObjectAttributesRequest

GetObjectAttributesResponse

GetObjectInformationRequest

GetObjectInformationResponse

GetSchemaAsJsonRequest

GetSchemaAsJsonResponse

GetTypedLinkFacetInformationRequest

GetTypedLinkFacetInformationResponse

IndexAttachment
Represents an index and an attached object.

LinkAttributeAction
The action to take on a typed link attribute value. Updates are only supported for attributes which don’t contribute to link identity.

LinkAttributeUpdate
Structure that contains attribute update information.

ListAppliedSchemaArnsRequest

ListAppliedSchemaArnsResponse

ListAttachedIndicesRequest

ListAttachedIndicesResponse

ListDevelopmentSchemaArnsRequest

ListDevelopmentSchemaArnsResponse

ListDirectoriesRequest

ListDirectoriesResponse

ListFacetAttributesRequest

ListFacetAttributesResponse

ListFacetNamesRequest

ListFacetNamesResponse

ListIncomingTypedLinksRequest

ListIncomingTypedLinksResponse

ListIndexRequest

ListIndexResponse

ListObjectAttributesRequest

ListObjectAttributesResponse

ListObjectChildrenRequest

ListObjectChildrenResponse

ListObjectParentPathsRequest

ListObjectParentPathsResponse

ListObjectParentsRequest

ListObjectParentsResponse

ListObjectPoliciesRequest

ListObjectPoliciesResponse

ListOutgoingTypedLinksRequest

ListOutgoingTypedLinksResponse

ListPolicyAttachmentsRequest

ListPolicyAttachmentsResponse

ListPublishedSchemaArnsRequest

ListPublishedSchemaArnsResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

ListTypedLinkFacetAttributesRequest

ListTypedLinkFacetAttributesResponse

ListTypedLinkFacetNamesRequest

ListTypedLinkFacetNamesResponse

LookupPolicyRequest

LookupPolicyResponse

ObjectAttributeAction
The action to take on the object attribute.

ObjectAttributeRange
A range of attributes.

ObjectAttributeUpdate
Structure that contains attribute update information.

ObjectReference
The reference that identifies an object.

PathToObjectIdentifiers
Returns the path to the ObjectIdentifiers that is associated with the directory.

PolicyAttachment
Contains the PolicyType, PolicyId, and the ObjectIdentifier to which it is attached. For more information, see Policies.

PolicyToPath
Used when a regular object exists in a Directory and you want to find all of the policies that are associated with that object and the parent to that object.

PublishSchemaRequest

PublishSchemaResponse

PutSchemaFromJsonRequest

PutSchemaFromJsonResponse

RemoveFacetFromObjectRequest

RemoveFacetFromObjectResponse

Rule
Contains an Amazon Resource Name (ARN) and parameters that are associated with the rule.

SchemaFacet
A facet.

Tag
The tag structure that contains a tag key and value.

TagResourceRequest

TagResourceResponse

TypedAttributeValue
Represents the data for a typed attribute. You can set one, and only one, of the elements. Each attribute in an item is a name-value pair. Attributes have a single value.

TypedAttributeValueRange
A range of attribute values. For more information, see Range Filters.

TypedLinkAttributeDefinition
A typed link attribute definition.

TypedLinkAttributeRange
Identifies the range of attributes that are used by a specified filter.

TypedLinkFacet
Defines the typed links structure and its attributes. To create a typed link facet, use the CreateTypedLinkFacet API.

TypedLinkFacetAttributeUpdate
A typed link facet attribute update.

TypedLinkSchemaAndFacetName
Identifies the schema Amazon Resource Name (ARN) and facet name for the typed link.

TypedLinkSpecifier
Contains all the information that is used to uniquely identify a typed link. The parameters discussed in this topic are used to uniquely specify the typed link being operated on. The AttachTypedLink API returns a typed link specifier while the DetachTypedLink API accepts one as input. Similarly, the ListIncomingTypedLinks and ListOutgoingTypedLinks API operations provide typed link specifiers as output. You can also construct a typed link specifier from scratch.

UntagResourceRequest

UntagResourceResponse

UpdateFacetRequest

UpdateFacetResponse

UpdateLinkAttributesRequest

UpdateLinkAttributesResponse

UpdateObjectAttributesRequest

UpdateObjectAttributesResponse

UpdateSchemaRequest

UpdateSchemaResponse

UpdateTypedLinkFacetRequest

UpdateTypedLinkFacetResponse

UpgradeAppliedSchemaRequest

UpgradeAppliedSchemaResponse

UpgradePublishedSchemaRequest

UpgradePublishedSchemaResponse

Enums

AddFacetToObjectError
Errors returned by AddFacetToObject

ApplySchemaError
Errors returned by ApplySchema

AttachObjectError
Errors returned by AttachObject

AttachPolicyError
Errors returned by AttachPolicy

AttachToIndexError
Errors returned by AttachToIndex

AttachTypedLinkError
Errors returned by AttachTypedLink

BatchReadError
Errors returned by BatchRead

BatchWriteError
Errors returned by BatchWrite

CreateDirectoryError
Errors returned by CreateDirectory

CreateFacetError
Errors returned by CreateFacet

CreateIndexError
Errors returned by CreateIndex

CreateObjectError
Errors returned by CreateObject

CreateSchemaError
Errors returned by CreateSchema

CreateTypedLinkFacetError
Errors returned by CreateTypedLinkFacet

DeleteDirectoryError
Errors returned by DeleteDirectory

DeleteFacetError
Errors returned by DeleteFacet

DeleteObjectError
Errors returned by DeleteObject

DeleteSchemaError
Errors returned by DeleteSchema

DeleteTypedLinkFacetError
Errors returned by DeleteTypedLinkFacet

DetachFromIndexError
Errors returned by DetachFromIndex

DetachObjectError
Errors returned by DetachObject

DetachPolicyError
Errors returned by DetachPolicy

DetachTypedLinkError
Errors returned by DetachTypedLink

DisableDirectoryError
Errors returned by DisableDirectory

EnableDirectoryError
Errors returned by EnableDirectory

GetAppliedSchemaVersionError
Errors returned by GetAppliedSchemaVersion

GetDirectoryError
Errors returned by GetDirectory

GetFacetError
Errors returned by GetFacet

GetLinkAttributesError
Errors returned by GetLinkAttributes

GetObjectAttributesError
Errors returned by GetObjectAttributes

GetObjectInformationError
Errors returned by GetObjectInformation

GetSchemaAsJsonError
Errors returned by GetSchemaAsJson

GetTypedLinkFacetInformationError
Errors returned by GetTypedLinkFacetInformation

ListAppliedSchemaArnsError
Errors returned by ListAppliedSchemaArns

ListAttachedIndicesError
Errors returned by ListAttachedIndices

ListDevelopmentSchemaArnsError
Errors returned by ListDevelopmentSchemaArns

ListDirectoriesError
Errors returned by ListDirectories

ListFacetAttributesError
Errors returned by ListFacetAttributes

ListFacetNamesError
Errors returned by ListFacetNames

ListIncomingTypedLinksError
Errors returned by ListIncomingTypedLinks

ListIndexError
Errors returned by ListIndex

ListObjectAttributesError
Errors returned by ListObjectAttributes

ListObjectChildrenError
Errors returned by ListObjectChildren

ListObjectParentPathsError
Errors returned by ListObjectParentPaths

ListObjectParentsError
Errors returned by ListObjectParents

ListObjectPoliciesError
Errors returned by ListObjectPolicies

ListOutgoingTypedLinksError
Errors returned by ListOutgoingTypedLinks

ListPolicyAttachmentsError
Errors returned by ListPolicyAttachments

ListPublishedSchemaArnsError
Errors returned by ListPublishedSchemaArns

ListTagsForResourceError
Errors returned by ListTagsForResource

ListTypedLinkFacetAttributesError
Errors returned by ListTypedLinkFacetAttributes

ListTypedLinkFacetNamesError
Errors returned by ListTypedLinkFacetNames

LookupPolicyError
Errors returned by LookupPolicy

PublishSchemaError
Errors returned by PublishSchema

PutSchemaFromJsonError
Errors returned by PutSchemaFromJson

RemoveFacetFromObjectError
Errors returned by RemoveFacetFromObject

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateFacetError
Errors returned by UpdateFacet

UpdateLinkAttributesError
Errors returned by UpdateLinkAttributes

UpdateObjectAttributesError
Errors returned by UpdateObjectAttributes

UpdateSchemaError
Errors returned by UpdateSchema

UpdateTypedLinkFacetError
Errors returned by UpdateTypedLinkFacet

UpgradeAppliedSchemaError
Errors returned by UpgradeAppliedSchema

UpgradePublishedSchemaError
Errors returned by UpgradePublishedSchema

Traits

CloudDirectory
Trait representing the capabilities of the Amazon CloudDirectory API. Amazon CloudDirectory clients implement this trait.